+++ {{cookiecutter.repo_name}}/project/mods_available/accounts/urls.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-
 urlpatterns = patterns('',
     url(r'^accounts/', include('project.accounts.urls')),
     ) + urlpatterns