+++ {{cookiecutter.repo_name}}/project/mods_available/porticus/settings.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-
 INSTALLED_APPS = add_to_tuple(INSTALLED_APPS, 'mptt', 'tagging', 'porticus', 'cmsplugin_porticus')
 
 PORTICUS_GALLERIES_PAGINATION = PORTICUS_ALBUMS_PAGINATION = PORTICUS_RESSOURCES_PAGINATION = 16