+++ {{cookiecutter.repo_name}}/project/mods_available/site_metas/settings.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-
-TEMPLATE_CONTEXT_PROCESSORS += (
+TEMPLATES[0]['OPTIONS']['context_processors'] = add_to_tuple(
+    TEMPLATES[0]['OPTIONS']['context_processors'],
     'project.utils.context_processors.site_metas',
 )