+++ {{cookiecutter.repo_name}}/project/mods_available/slideshows/settings.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-
 INSTALLED_APPS = add_to_tuple(INSTALLED_APPS, 
     'slideshows',
     'cmsplugin_slideshows',