+++ {{cookiecutter.repo_name}}/project/mods_available/thumbnails/settings.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 INSTALLED_APPS = add_to_tuple(INSTALLED_APPS, 'easy_thumbnails')
 
 # easy_thumbnails for django_filer